lesson4 RoomService code improve

+++ src/main/java/lesson4/service/RoomService.java
@@ -4,16 +4,15 @@ import lesson4.DAO.HotelDAO;
 import lesson4.DAO.RoomDAO;
 import lesson4.exception.*;
 import lesson4.model.Filter;
-import lesson4.model.Hotel;
 import lesson4.model.Room;
 
 import java.util.List;
 
 public class RoomService {
 
+    private static final UserService userService = new UserService();
     private static final RoomDAO roomDAO = new RoomDAO();
     private static final HotelDAO hotelDAO = new HotelDAO();
-    private static final UserService userService = new UserService();
 
     public List<Room> findRooms(Filter filter) throws InternalServerException, BadRequestException, NotFoundException {
         validateFilter(filter);
@@ -21,41 +20,51 @@ public class RoomService {
     }
 
     public Room addRoom(Room room)
-            throws InternalServerException, BadRequestException, NoAccessException, NotFoundException,
-            NotLogInException {
+            throws InternalServerException, BadRequestException, NoAccessException, NotLogInException {
+
         userService.checkForAdminPermissions();
         validateRoom(room);
         return roomDAO.save(room);
     }
 
-    public void deleteRoom(long roomId) throws InternalServerException, NoAccessException, NotFoundException,
-            NotLogInException {
+    public void deleteRoom(long roomId)
+            throws InternalServerException, NoAccessException, NotLogInException, BadRequestException {
+
         userService.checkForAdminPermissions();
-        Room room = roomDAO.findById(roomId);
+
+        if (!roomDAO.isExist(roomId)) {
+            throw new BadRequestException("Missing room with id: " + roomId);
+        }
+        Room room = new Room();
+        room.setId(roomId);
         roomDAO.delete(room);
     }
 
     private void validateFilter(Filter filter) throws BadRequestException {
-        if (filter == null ||
-                (filter.getNumberOfGuests() == 0 && filter.getPrice() == 0 &&
-                        filter.getBreakfastIncluded() == null && filter.getPetsAllowed() == null &&
-                        filter.getDateAvailableFrom() == null && filter.getCountry() == null &&
-                        filter.getCity() == null)) {
-            throw new BadRequestException("validateFilter failed: you have not selected any options for filtering");
-        }
-        if (filter.getNumberOfGuests() < 0 && filter.getPrice() < 0) {
-            throw new BadRequestException("validateFilter failed: you have not selected correct options for filtering");
+        if (filter.getNumberOfGuests() != null
+                    && filter.getNumberOfGuests() < 0
+                || filter.getPrice() != null
+                    && filter.getPrice() < 0) {
+            throw new BadRequestException("One of numeric options is incorrect");
         }
     }
 
-    private void validateRoom(Room room) throws InternalServerException, BadRequestException, NotFoundException {
+    private void validateRoom(Room room) throws InternalServerException, BadRequestException {
         if (room == null) {
-            throw new BadRequestException("validateRoom failed: impossible to process null room");
+            throw new BadRequestException("Impossible to process null room");
         }
-        if (room.getNumberOfGuests() <= 0 || room.getPrice() <= 0.0 || room.isBreakfastIncluded() == null ||
-                room.isPetsAllowed() == null || room.getDateAvailableFrom() == null || room.getHotel() == null) {
-            throw new BadRequestException("validateRoom failed: not all fields are filled correctly");
+        if (room.getNumberOfGuests() == null
+                || room.getPrice() == null
+                || room.getBreakfastIncluded() == null
+                || room.getPetsAllowed() == null
+                || room.getDateAvailableFrom() == null
+                || room.getHotel() == null
+                || room.getNumberOfGuests() < 1
+                || room.getPrice() <= 0.0) {
+            throw new BadRequestException("Not all fields are filled correctly");
         }
-        hotelDAO.findById(room.getHotel().getId());
+        if (!hotelDAO.isExist(room.getHotel().getId())) {
+            throw new BadRequestException("Missing hotel with id: " + room.getHotel().getId());
+        }
     }
 }
